refactor: Merge Data Model Interfaces
