+++ common/changes/@yuants/agent/2023-12-21-11-04.json
@@ -1,0 +1,10 @@
+{
+  "changes": [
+    {
+      "packageName": "@yuants/agent",
+      "comment": "upgrade ",
+      "type": "patch"
+    }
+  ],
+  "packageName": "@yuants/agent"
+}
