+++ common/changes/@yuants/protocol/2023-12-21-11-04.json
@@ -1,0 +1,10 @@
+{
+  "changes": [
+    {
+      "packageName": "@yuants/protocol",
+      "comment": "merge model interfaces with data-model package",
+      "type": "minor"
+    }
+  ],
+  "packageName": "@yuants/protocol"
+}
