+++ libraries/data-model/src/utils.ts
@@ -13,6 +13,7 @@ export const createEmptyAccountInfo = (
   leverage: number = 1,
   initial_balance: number = 0,
 ): IAccountInfo => ({
+  timestamp_in_us: 0,
   updated_at: 0,
   account_id,
   money: {
