+++ libraries/protocol/etc/protocol.api.md
@@ -4,10 +4,20 @@
 
 ```ts
 
+import { IAccountInfo } from '@yuants/data-model';
 import { IAccountMoney } from '@yuants/data-model';
 import { IDataRecord } from '@yuants/data-model';
+import { IOrder } from '@yuants/data-model';
+import { IPeriod } from '@yuants/data-model';
+import { IPosition } from '@yuants/data-model';
+import { IProduct } from '@yuants/data-model';
+import { ITick } from '@yuants/data-model';
 import { JSONSchema7 } from 'json-schema';
 import { Observable } from 'rxjs';
+import { OrderDirection } from '@yuants/data-model';
+import { OrderStatus } from '@yuants/data-model';
+import { OrderType } from '@yuants/data-model';
+import { PositionVariant } from '@yuants/data-model';
 import { Registry } from '@yuants/prometheus-client';
 import { Subject } from 'rxjs';
 
@@ -17,14 +27,7 @@ export function createConnectionJson<T = any>(URL: string): IConnection<T>;
 // @public
 export function createConnectionWs<T = any>(URL: string): IConnection<T>;
 
-// @public
-export interface IAccountInfo {
-    account_id: string;
-    money: IAccountMoney;
-    orders: IOrder[];
-    positions: IPosition[];
-    timestamp_in_us: number;
-}
+export { IAccountInfo }
 
 export { IAccountMoney }
 
@@ -37,78 +40,13 @@ export interface IConnection<T> {
 
 export { IDataRecord }
 
-// @public
-export interface IOrder {
-    account_id: string;
-    client_order_id: string;
-    comment?: string;
-    direction: OrderDirection;
-    exchange_order_id?: string;
-    inferred_base_currency_price?: number;
-    originator?: string;
-    position_id?: string;
-    price?: number;
-    product_id: string;
-    profit_correction?: number;
-    real_profit?: number;
-    status?: OrderStatus;
-    stop_loss_price?: number;
-    take_profit_price?: number;
-    timestamp_in_us?: number;
-    traded_price?: number;
-    traded_volume?: number;
-    type: OrderType;
-    volume: number;
-}
+export { IOrder }
 
-// @public
-export interface IPeriod {
-    close: number;
-    datasource_id: string;
-    high: number;
-    low: number;
-    open: number;
-    open_interest?: number;
-    period_in_sec: number;
-    product_id: string;
-    spread?: number;
-    timestamp_in_us: number;
-    volume: number;
-}
+export { IPeriod }
 
-// @public
-export interface IPosition {
-    closable_price: number;
-    comment?: string;
-    floating_profit: number;
-    free_volume: number;
-    position_id: string;
-    position_price: number;
-    product_id: string;
-    variant: PositionVariant;
-    volume: number;
-}
+export { IPosition }
 
-// @public
-export interface IProduct {
-    allow_long?: boolean;
-    allow_short?: boolean;
-    base_currency: string;
-    datasource_id: string;
-    is_underlying_base_currency?: boolean;
-    margin_rate?: number;
-    max_position?: number;
-    max_volume?: number;
-    name?: string;
-    price_step?: number;
-    product_id: string;
-    quoted_currency?: string;
-    spread?: number;
-    value_based_cost?: number;
-    value_speed?: number;
-    volume_based_cost?: number;
-    volume_step?: number;
-}
+export { IProduct }
 
 // @public
 export interface IResponse<T = void> {
@@ -171,48 +109,15 @@ export interface ITerminalMessage {
     trace_id: string;
 }
 
-// @public
-export interface ITick {
-    ask?: number;
-    bid?: number;
-    datasource_id: string;
-    open_interest?: number;
-    price: number;
-    product_id: string;
-    spread?: number;
-    timestamp_in_us: number;
-    volume: number;
-}
+export { ITick }
 
-// @public
-export enum OrderDirection {
-    CLOSE_LONG = 1,
-    CLOSE_SHORT = 3,
-    OPEN_LONG = 0,
-    OPEN_SHORT = 2
-}
+export { OrderDirection }
 
-// @public
-export enum OrderStatus {
-    ACCEPTED = 0,
-    CANCELLED = 2,
-    TRADED = 1
-}
+export { OrderStatus }
 
-// @public
-export enum OrderType {
-    FOK = 3,
-    IOC = 4,
-    LIMIT = 1,
-    MARKET = 0,
-    STOP = 2
-}
+export { OrderType }
 
-// @public
-export enum PositionVariant {
-    LONG = 0,
-    SHORT = 1
-}
+export { PositionVariant }
 
 // @public
 export const PromRegistry: Registry;
