+++ libraries/protocol/src/model.ts
@@ -1,644 +1,18 @@
-import { IAccountMoney } from '@yuants/data-model';
 import { JSONSchema7 } from 'json-schema';
-export { IAccountMoney, IDataRecord } from '@yuants/data-model';
-
-/**
- * Order: Changes the Position of the AccountInfo in the account through a trading command.
- * 订单: 通过交易命令改变账户内头寸
- * @public
- */
-export interface IOrder {
-  /**
-   * Client order ID.
-   * 客户端订单ID
-   */
-  client_order_id: string;
-  /**
-   * Exchange order ID (if any).
-   * 交易所订单ID (如果有)
-   */
-  exchange_order_id?: string;
-  /**
-   * Order source specified by the order placer.
-   * 下单器指定的订单来源
-   */
-  originator?: string;
-
-  /**
-   * Account ID.
-   * 账户 ID
-   */
-  account_id: string;
-  /**
-   * Product ID.
-   * 品种 ID
-   */
-  product_id: string;
-
-  /**
-   * Specifies the position ID to be operated on.
-   * 指定需要操作的头寸 ID
-   *
-   * - If left blank, it means "any position under the same account and product can be operated on".
-   * - 如果留空，表达 "操作同账户、同品种下哪个具体头寸都可以"
-   * - If filled in, only the matching position can be operated on, and other positions under the same account and product cannot be affected.
-   * - 如果填写了，只能操作匹配的头寸，不得影响同账户下同品种下的其他头寸。
-   *
-   */
-  position_id?: string;
-
-  /**
-   * Order type.
-   * 订单类型
-   */
-  type: OrderType;
-  /**
-   * Order direction.
-   * 订单方向
-   */
-  direction: OrderDirection;
-  /**
-   * Order volume.
-   * 委托量
-   */
-  volume: number;
-
-  /**
-   * Profit and loss correction in non-standard models.
-   * 非标准模型下的盈亏修正
-   *
-   * When the profit and loss model is non-standard,
-   * a profit and loss correction value can be added to correct the standard model profit and loss to the actual profit and loss.
-   * 当盈亏模型非标准时，可以添加一个盈亏修正值，将标准模型盈亏修正到实际盈亏。
-   *
-   * Profit and loss correction = actual profit and loss - standard profit and loss
-   * 盈亏修正 = 实际盈亏 - 标准盈亏
-   *
-   * If this value is empty, it is semantically equivalent to 0.
-   * 如果此值为空，语义等同于 0
-   *
-   * See [How to Calculate Profit and Loss](https://tradelife.feishu.cn/wiki/wikcnRNzWSF7jtkH8nGruaMhhlh) for reference.
-   * 参考 [如何计算盈亏](https://tradelife.feishu.cn/wiki/wikcnRNzWSF7jtkH8nGruaMhhlh)
-   *
-   */
-  profit_correction?: number;
-
-  /**
-   * Actual profit and loss.
-   * 实际盈亏
-   *
-   * The amount of change in the account balance when closing a position.
-   * 平仓时，对账户的余额产生的改变量
-   *
-   * If this value is empty, it is semantically equivalent to "profit_correction == 0", i.e., "standard profit and loss == actual profit and loss".
-   * 如果此值为空，语义等同于 "盈亏修正 == 0" 即 "标准盈亏 == 实际盈亏"
-   *
-   * See [How to Calculate Profit and Loss](https://tradelife.feishu.cn/wiki/wikcnRNzWSF7jtkH8nGruaMhhlh) for reference.
-   * 参考 [如何计算盈亏](https://tradelife.feishu.cn/wiki/wikcnRNzWSF7jtkH8nGruaMhhlh)
-   */
-  real_profit?: number;
-
-  /**
-   * The inferred price of the base currency against the margin currency at the time of closing the position.
-   * 推断得到的平仓时基准货币兑保证金货币的价格
-   *
-   * If this value is empty, it is semantically equivalent to 1 (i.e., the base currency is the same as the margin currency).
-   * 如果此值为空，语义等同于 1 (即基准货币 == 保证金货币)
-   *
-   * See [How to Calculate Profit and Loss](https://tradelife.feishu.cn/wiki/wikcnRNzWSF7jtkH8nGruaMhhlh) for reference.
-   * 参考 [如何计算盈亏](https://tradelife.feishu.cn/wiki/wikcnRNzWSF7jtkH8nGruaMhhlh)
-   */
-  inferred_base_currency_price?: number;
-
-  /**
-   * Order timestamp / trade timestamp.
-   * 下单时间戳 / 成交时间戳
-   */
-  timestamp_in_us?: number;
-  /**
-   * Order price.
-   * 委托价
-   */
-  price?: number;
-  /**
-   * Traded volume.
-   * 成交量
-   */
-  traded_volume?: number;
-  /**
-   * Traded price.
-   * 成交价
-   */
-  traded_price?: number;
-  /**
-   * Order status.
-   * 订单状态
-   */
-  status?: OrderStatus;
-  /**
-   * Order comment.
-   * 订单注释
-   */
-  comment?: string;
-  /**
-   * Take profit price (ignored for now).
-   * 止盈价 (暂时不可用)
-   */
-  take_profit_price?: number;
-  /**
-   * Stop loss price (ignored for now).
-   * 止损价 (暂时不可用)
-   */
-  stop_loss_price?: number;
-}
-/**
- * Position variant.
- * 头寸类型
- * @public
- */
-
-export enum PositionVariant {
-  /**
-   * Long position.
-   * 做多
-   */
-  LONG = 0,
-  /**
-   * Short position.
-   * 做空
-   */
-  SHORT = 1,
-}
-
-/**
- * Position: Atomic position information.
- * 原子性的持仓头寸信息
- *
- * Positions on the same product can be aggregated.
- * 相同品种上的头寸可以被合计
- *
- * @public
- */
-export interface IPosition {
-  /**
-   * Position ID.
-   * 头寸 ID
-   */
-  position_id: string;
-  /**
-   * Product ID.
-   * 品种 ID
-   */
-  product_id: string;
-  /**
-   * Position variant.
-   * 仓位类型
-   *
-   * can be used to calculate net position according to position type
-   * 可以根据仓位类型计算净头寸
-   */
-  variant: PositionVariant;
-  /**
-   * Position volume (non-negative).
-   * 持仓量 (非负)
-   *
-   * When calculating net value, this field should be referenced.
-   * 结算净值时应参考此字段
-   */
-  volume: number;
-  /**
-   * Tradable volume (non-negative).
-   * 可交易量 (非负)
-   *
-   * When placing an order, this field should be referenced.
-   * 下单时应检查此字段
-   *
-   * For T+0 trading, this field should be consistent with the volume field;
-   * 市场为 T+0 交易时应当与 volume 字段一致;
-   * For T+1 trading, this field may be smaller than the volume field.
-   * 市场为 T+1 交易时，可能比 volume 小.
-   */
-  free_volume: number;
-  /**
-   * Position price.
-   * 持仓成本价 (可通过 product_id 得到价格的内在含义)
-   */
-  position_price: number;
-
-  /**
-   * The current closable settlement price.
-   * 当前可平仓结算价格
-   */
-  closable_price: number;
-
-  /**
-   * Floating profit and loss of the position.
-   * 持仓浮动盈亏
-   */
-  floating_profit: number;
-
-  /**
-   * the comment of the position.
-   * 头寸的备注
-   */
-  comment?: string;
-
-  // Position is one of the reasons for occupying margin,
-  // 持仓是造成占用保证金的原因之一
-  // but the calculation mechanism of margin is relatively complex, and the algorithms of various exchanges are different.
-  // 但保证金的计算机制比较复杂，各交易所算法不一
-  // Therefore, Yuan does not calculate the margin based on the final margin given by the exchange.
-  // 因此以交易所给出的最终保证金为准，Yuan 不进行保证金的计算
-  // margin: number;
-}
-
-/** 账户信息 @public */
-export interface IAccountInfo {
-  /**
-   * Account ID.
-   * 账户ID
-   */
-  account_id: string;
-  /**
-   * Money information.
-   * 资金信息
-   */
-  money: IAccountMoney;
-  /**
-   * Position information.
-   * 持仓信息
-   */
-  positions: IPosition[];
-  /**
-   * Unfilled orders.
-   * 未成交的挂单
-   */
-  orders: IOrder[];
-  /**
-   * Timestamp when the account information was generated.
-   * 账户信息产生的时间戳
-   *
-   * (Used to handle conflicts: always accept the latest information)
-   * (用于处理冲突: 应当总是接受最新的信息)
-   */
-  timestamp_in_us: number;
-}
-
-/**
- * 品种: 交易的标的物
- *
- * @public
- */
-export interface IProduct {
-  /** Data source ID */
-  datasource_id: string;
-  /** Product ID */
-  product_id: string;
-  /** Readable product name */
-  name?: string;
-
-  /**
-   * Base Currency
-   * 基准货币 (Base Currency)
-   *
-   * The base currency is the currency used as the basis for exchange rate quotes, expressed as the number of units of the currency that can be exchanged for one unit of the quoted currency.
-   * 基准货币是汇率报价中作为基础的货币，即报价表达形式为每一个单位的货币可兑换多少另一种货币。
-   *
-   * e.g. The base currency of GBPJPY is GBP; the base currency of USDCAD is USD.
-   * e.g. GBPJPY 的 base_currency 为 GBP; USDCAD 的 base_currency 为 USD.
-   */
-  base_currency: string;
-
-  /**
-   * Quoted Currency
-   * 标价货币 (Quoted Currency)
-   *
-   * The quoted currency is the currency being used as the reference for the exchange rate quote, expressed as the number of units of the quoted currency that can be exchanged for one unit of the base currency.
-   * 汇率的表达方式为一单位的基准货币可兑换多少单位的标价货币
-   *
-   * e.g. The quoted currency of GBPJPY is JPY; the quoted currency of USDCAD is CAD.
-   * e.g. GBPJPY 的 quoted_currency 为 JPY; USDCAD 的 quoted_currency 为 CAD.
-   *
-   * For non-forex products, the quoted currency should be empty.
-   * 对于非外汇品种，quoted_currency 应当为空。
-   */
-  quoted_currency?: string;
-
-  /**
-   * Is the underlying asset the base currency?
-   * 标的物是基准货币吗？
-   *
-   * One lot corresponds to the quantity of the underlying asset specified by value_speed, which can be the base currency or other commodities.
-   * 1 手对应 value_speed 数量的标的物，此标的物可以是基准货币或其他商品。
-   *
-   * - For commodities, including spot and futures, this value is usually false, because one lot corresponds to a multiple of value_speed of the commodity quantity.
-   * - 对于商品，包括现货和期货，此值通常为 false，因为 1 手对应着 value_speed 倍数的商品数量。
-   *
-   * - For forex, this value is usually true, because one lot corresponds to a multiple of value_speed of the equivalent of the base currency in any currency.
-   * - 对于外汇，此值通常为 true，因为 1 手对应着 value_speed 倍数的基础货币等值的任一货币。
-   *
-   * If the value is empty, it is semantically equivalent to false.
-   * 如果值为空，语义上等同于 false.
-   *
-   * If this value is true, an additional division by the "closing price" of this product is required in the standard yield formula.
-   * 如果此值为 true，需要在标准收益公式中额外除以本品种的"平仓时的价格"。
-   */
-  is_underlying_base_currency?: boolean;
-
-  /**
-   * price step, default is 1
-   * 报价单位，默认为 1
-   */
-  price_step?: number;
-  /**
-   * Volume unit (unit: lot), default is 1
-   * 成交量单位 (单位: 手)，默认为 1
-   */
-  volume_step?: number;
-  /**
-   * Value speed, default is 1
-   * 价值速率，默认为 1
-   *
-   * The quantity of the underlying asset specified by one lot.
-   * 1 手对应的标的物的数量
-   *
-   * ~~For every 1 lot increase in price, the settlement asset income obtained~~
-   * ~~每做多 1 手，价格每上升 1，获得的结算资产收益~~
-   */
-  value_speed?: number;
-
-  /**
-   * Margin rate
-   * 保证金率
-   *
-   * Margin calculation reference [How to calculate margin](https://tradelife.feishu.cn/wiki/wikcnEVBM0RQ7pmbNZUxMV8viRg)
-   * 保证金计算参考 [如何计算保证金](https://tradelife.feishu.cn/wiki/wikcnEVBM0RQ7pmbNZUxMV8viRg)
-   */
-  margin_rate?: number;
-
-  /**
-   * Value-based cost
-   * 基于价值的成本
-   */
-  value_based_cost?: number;
-  /**
-   * Volume-based cost
-   * 基于成交量的成本
-   */
-  volume_based_cost?: number;
-
-  /**
-   * Maximum position
-   * 最大持仓量
-   */
-  max_position?: number;
-  /** 最大单笔委托量 */
-  max_volume?: number;
-
-  /**
-   * Allow long
-   * 允许做多
-   *
-   * If this value is empty, it is semantically equivalent to true.
-   * 如果此值为空，语义上等同于 true.
-   */
-  allow_long?: boolean;
-  /**
-   * Allow short
-   * 允许做空
-   *
-   * If this value is empty, it is semantically equivalent to true.
-   * 如果此值为空，语义上等同于 true.
-   */
-  allow_short?: boolean;
-
-  /**
-   * Spread
-   * 预期点差
-   */
-  spread?: number;
-}
-/**
- * 订单类型
- * @public
- */
-
-/**
- * 订单类型
- * Order Type
- * @public
- */
-
-export enum OrderType {
-  /**
-   * Market Order: Executed at the current market price
-   * 市价单: 以市场价格成交
-   *
-   * The most common and simple order type, no need to specify an order price
-   * 最普遍而简单的订单类型，不需要指定委托价
-   */
-  MARKET = 0,
-  /**
-   * Limit Order: Limits the price at which the order can be executed
-   * 限价单: 限制成交的价格
-   *
-   * - BUY LIMIT: The execution price will not be higher than the order price
-   * - SELL LIMIT: The execution price will not be lower than the order price
-   *
-   * - BUY LIMIT: 成交价不会高于委托价
-   * - SELL LIMIT: 成交价不会低于委托价
-   */
-  LIMIT = 1,
-  /**
-   * Stop Order: Triggers a market order when the market price reaches the order price
-   * 触发单: 市场价达到委托价时触发市价单
-   *
-   * - BUY STOP: Place an order when the market price is higher than the order price
-   * - SELL STOP: Place an order when the market price is lower than the order price
-   *
-   * - BUY STOP: 市场价高于委托价时下单
-   * - SELL STOP: 市场价低于委托价时下单
-   */
-  STOP = 2,
-  /**
-   * Fill or Kill: Requires immediate and complete
-   * 即成或撤单: Fill or Kill
-   *
-   * It is required to be executed immediately and completely when placing an order, otherwise it will be cancelled
-   * 下单时要求立即全部成交，否则撤单
-   */
-  FOK = 3,
-  /**
-   * Immediate or Cancel: Requires immediate execution, allows partial execution, and cancels the rest
-   * 即成余撤单: Immediate or Cancel
-   *
-   * It is required to be executed immediately when placing an order, allows partial execution, and cancels the rest
-   * 下单时要求立即成交，允许部分成交，未成交的直接撤单
-   */
-  IOC = 4,
-}
-/**
- * 订单方向
- * @public
- */
-
-export enum OrderDirection {
-  /**
-   * Open long position
-   * 开多
-   */
-  OPEN_LONG = 0,
-  /**
-   * Close long position
-   * 平多
-   */
-  CLOSE_LONG = 1,
-  /**
-   * Open short position
-   * 开空
-   */
-  OPEN_SHORT = 2,
-  /**
-   * Close short position
-   * 平空
-   */
-  CLOSE_SHORT = 3,
-}
-/**
- * 订单状态
- * @public
- */
-
-export enum OrderStatus {
-  /**
-   * Order accepted by the exchange
-   * 交易所已接受委托
-   */
-  ACCEPTED = 0,
-  /**
-   * Order partially filled
-   * 已成交
-   */
-  TRADED = 1,
-  /**
-   * Order cancelled
-   * 已撤单
-   */
-  CANCELLED = 2,
-}
-/**
- * Tick: Market transaction data at a certain moment
- * Tick: 某个时刻的市场成交行情数据
- * @public
- */
-
-export interface ITick {
-  /**
-   * Data source ID
-   * 数据源 ID
-   */
-  datasource_id: string;
-  /**
-   * Product ID
-   * 品种 ID
-   */
-  product_id: string;
-  /**
-   * Timestamp (in microseconds)
-   * 时间戳
-   */
-  timestamp_in_us: number;
-  /**
-   * Price
-   * 成交价
-   */
-  price: number;
-  /**
-   * Volume
-   * 成交量
-   */
-  volume: number;
-  /**
-   * Open interest
-   * 持仓量
-   */
-  open_interest?: number;
-  /**
-   * Spread
-   * 点差
-   */
-  spread?: number;
-  /**
-   * Ask price
-   * 卖一价
-   */
-  ask?: number;
-  /** 买一价 */
-  bid?: number;
-}
-/**
- * Period: Market transaction data during a certain period of time
- * Period: 某个时间段内的市场成交行情数据
- * @public
- */
-
-export interface IPeriod {
-  /**
-   * Data source ID
-   * 数据源 ID
-   */
-  datasource_id: string;
-  /**
-   * Product ID
-   * 品种 ID
-   */
-  product_id: string;
-  /**
-   * Period (in seconds)
-   * 时间周期 (秒)
-   */
-  period_in_sec: number;
-  /**
-   * Start timestamp (open, in microseconds)
-   * 开始时间戳 (open)
-   */
-  timestamp_in_us: number;
-  /**
-   * Opening price
-   * 开盘价
-   */
-  open: number;
-  /**
-   * Highest price
-   * 最高价
-   */
-  high: number;
-  /**
-   * Lowest price
-   * 最低价
-   */
-  low: number;
-  /**
-   * Closing price
-   * 收盘价
-   */
-  close: number;
-  /**
-   * Volume
-   * 成交量
-   */
-  volume: number;
-  /**
-   * Open interest
-   * 持仓量
-   */
-  open_interest?: number;
-  /**
-   * Spread
-   * 点差
-   */
-  spread?: number;
-}
+export {
+  IAccountInfo,
+  IAccountMoney,
+  IDataRecord,
+  IOrder,
+  IPeriod,
+  IPosition,
+  IProduct,
+  ITick,
+  OrderDirection,
+  OrderStatus,
+  OrderType,
+  PositionVariant,
+} from '@yuants/data-model';
 
 /**
  * 终端基本信息
